add edit cash

+++ PublicAPI/Endpoints/Portfolio/Delete.cs
@@ -2,6 +2,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using ApplicationCore.PortfolioAggregate;
 using Mapster;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace PublicAPI.Endpoints.Portfolio
@@ -9,15 +10,22 @@ namespace PublicAPI.Endpoints.Portfolio
     public class Delete: BasePortfolioRelatedEndpoint<int, PortfolioResponse>
     {
         private readonly IPortfolioService _portfolioService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public Delete(IPortfolioService portfolioService)
+        public Delete(IPortfolioService portfolioService, IAuthorizationService authorizationService)
         {
             _portfolioService = portfolioService;
+            _authorizationService = authorizationService;
         }
-
+        
+        [HttpDelete]
         public override async  Task<ActionResult<PortfolioResponse>> 
             HandleAsync(int portfolioId, CancellationToken cancellationToken = new CancellationToken())
         {
+            if (!await IsAllowedToExecute(portfolioId, _authorizationService))
+            {
+                return Unauthorized(); 
+            }
             var result = _portfolioService.DeletePortfolio(portfolioId);
             if (result is null)
                 return NotFound();
